AJAXServlet is the global servlet handling all AJAX request coming from the client. Now the functionality of the FormConfigurationLoader is moved to the AJAXServlet. Accordingly the servlet mappings are changed in the web.xml.
